[Feat] 캐릭터 공격에 따른 데미지 표시, 치명타 적용

+++ Assets/PersonalWorkplace/LGH/Scripts/PlayerModel.cs
@@ -18,7 +18,7 @@ public class PlayerModel : MonoBehaviour
     public float skillDamageBonus = 1;
 
     // 총 전투력
-    public double combatPower;
+    public float combatPower;
 
     // Observable Properties
     public ObservableProperty<double> CurHealth { get; private set; } = new();
